requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
langchain==0.1.0
langchain-openai==0.0.2
chromadb==0.4.18
pypdf==3.17.1
python-multipart==0.0.6
python-dotenv==1.0.0
openai>=1.6.1,<2.0.0
sentence-transformers>=2.3.0
tiktoken==0.5.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov
